+++ packages/leptos/input/src/default.rs
@@ -5,8 +5,41 @@ use leptos_style::Style;
 use shadcn_ui_leptos_utils::generate_handler;
 use tailwind_fuse::*;
 
-use crate::types::InputType;
+#[derive(Default, Clone)]
+pub enum InputType {
+    #[default]
+    Text,
+    Password,
+    Search,
+    Tel,
+    Url,
+    Email,
+    Time,
+    Date,
+    DatetimeLocal,
+    Month,
+    Week,
+    File,
+}
 
+impl InputType {
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            Self::Text => "text",
+            Self::Password => "password",
+            Self::Search => "search",
+            Self::Tel => "tel",
+            Self::Url => "url",
+            Self::Email => "email",
+            Self::Time => "time",
+            Self::Date => "date",
+            Self::DatetimeLocal => "datetime-local",
+            Self::Month => "month",
+            Self::Week => "week",
+            Self::File => "file",
+        }
+    }
+}
 
 #[component]
 pub fn Input(
