--- packages/leptos/input/src/types.rs
@@ -1,33 +1,0 @@
-#[derive(Default, Clone)]
-pub enum InputType {
-    #[default]
-    Text,
-    Password,
-    Search,
-    Tel,
-    Url,
-    Email,
-    Time,
-    Date,
-    DatetimeLocal,
-    Month,
-    Week,
-}
-
-impl InputType {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Self::Text => "text",
-            Self::Password => "password",
-            Self::Search => "search",
-            Self::Tel => "tel",
-            Self::Url => "url",
-            Self::Email => "email",
-            Self::Time => "time",
-            Self::Date => "date",
-            Self::DatetimeLocal => "datetime-local",
-            Self::Month => "month",
-            Self::Week => "week",
-        }
-    }
-}
